api/requirements.txt
typing-extensions==4.4.0
SQLAlchemy==1.4.46
pydantic==1.10.4
fastapi==0.89.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5